Split Live Jobs into own component
